Refactored and added keycloak
