+++ .env
@@ -6,3 +6,5 @@ ORDER_DB_USER=order_service
 ORDER_DB_PASSWORD=order_service
 RABBITMQ_USER=rabbitmq
 RABBITMQ_PASSWORD=rabbitmq_pass
+KEYCLOAK_USER=keycloak_password
+KEYCLOAK_PASSWORD=keycloak_password
