+++ config/src/main/resources/shared/gateway.yml
@@ -1,3 +1,6 @@
+server:
+  port: 9000
+
 zuul:
   ignoredServices: '*'
