--- config/src/main/resources/shared/shopping-cart-service.yml
@@ -1,13 +1,0 @@
-app:
-  service:
-    shopping-cart:
-      cookie:
-        ageHours: 5040
-
-feign:
-  hystrix:
-    enabled: true
-
-logging:
-  level:
-    com.mikhailkarpov.eshop.shoppingcartservice.client.ProductServiceClient: DEBUG
