+++ docker-compose.yml
@@ -6,16 +6,19 @@ services:
     image: dadarek/wait-for-dependencies
     depends_on:
       - config
+      - keycloak
       - product-service-db
       - order-service-db
       - rabbitmq
       - redis
-    command: config:8888 product-service-db:5432 order-service-db:5432 rabbitmq:5672 redis:6379
+
+    command: config:8888 product-service-db:5432 order-service-db:5432 rabbitmq:5672 redis:6379 keycloak:8080
 
   config:
     build: ./config
     container_name: config
     environment:
+      - CONFIG_USER=${CONFIG_USER}
       - CONFIG_PASSWORD=${CONFIG_PASSWORD}
     ports:
     - 8888:8888
@@ -27,6 +30,7 @@ services:
       - config
     environment:
       - CONFIG_URI=http://config:8888
+      - CONFIG_USER=${CONFIG_USER}
       - CONFIG_PASSWORD=${CONFIG_PASSWORD}
       - SPRING_PROFILES_ACTIVE=dev
     ports:
@@ -39,14 +43,24 @@ services:
       - config
     environment:
       - CONFIG_URI=http://config:8888
+      - CONFIG_USER=${CONFIG_USER}
       - CONFIG_PASSWORD=${CONFIG_PASSWORD}
       - SPRING_PROFILES_ACTIVE=dev
+    ports:
+      - 9000:8080
+
+  keycloak:
+    image: jboss/keycloak
+    container_name: keycloak
+    environment:
+      - KEYCLOAK_USER={KEYCLOAK_USER}
+      - KEYCLOAK_PASSWORD={KEYCLOAK_PASSWORD}
     ports:
       - 8080:8080
 
   product-service:
     build: ./product-service
-    container_name: product-service-db
+    container_name: product-service
     depends_on:
       - config
       - product-service-db
@@ -54,6 +68,7 @@ services:
       - redis
     environment:
       - CONFIG_URI=http://config:8888
+      - CONFIG_USER=${CONFIG_USER}
       - CONFIG_PASSWORD=${CONFIG_PASSWORD}
       - SPRING_PROFILES_ACTIVE=dev
       - PRODUCT_DB_USER=${PRODUCT_DB_USER}
@@ -62,7 +77,7 @@ services:
       - RABBITMQ_PASSWORD=${RABBITMQ_PASSWORD}
 
   product-service-db:
-    image: "postgres:12"
+    image: "postgres"
     container_name: product-serivce-db
     environment:
       - POSTGRES_DB=product_service
@@ -80,6 +95,7 @@ services:
       - order-service-db
     environment:
       - CONFIG_URI=http://config:8888
+      - CONFIG_USER=${CONFIG_USER}
       - CONFIG_PASSWORD=${CONFIG_PASSWORD}
       - SPRING_PROFILES_ACTIVE=dev
       - ORDER_DB_USER=${ORDER_DB_USER}
@@ -88,7 +104,7 @@ services:
       - RABBITMQ_PASSWORD=${RABBITMQ_PASSWORD}
 
   order-service-db:
-    image: "postgres:12"
+    image: "postgres"
     container_name: order-service-db
     environment:
       - POSTGRES_DB=order_service
@@ -108,7 +124,7 @@ services:
     - 15672:15672
 
   redis:
-    image: redis:alpine
+    image: redis
     container_name: redis
     ports:
     - 6379:6379
