+++ gateway/src/main/resources/bootstrap.yml
@@ -6,5 +6,5 @@ spring:
     config:
       uri: ${CONFIG_URI:http://localhost:8888}
       fail-fast: true
-      username: config
+      username: ${CONFIG_USER}
       password: ${CONFIG_PASSWORD}
