+++ order-service/src/test/java/com/mikhailkarpov/eshop/orders/OrderServiceApplicationIT.java
@@ -1,6 +1,5 @@
 package com.mikhailkarpov.eshop.orders;
 
-import com.mikhailkarpov.eshop.orders.config.AbstractIT;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
