+++ product-service/src/main/resources/bootstrap.yml
@@ -9,5 +9,5 @@ spring:
     config:
       uri: ${CONFIG_URI:http://localhost:8888}
       fail-fast: true
-      username: config
+      username: ${CONFIG_USER}
       password: ${CONFIG_PASSWORD}
