+++ product-service/src/test/java/com/mikhailkarpov/eshop/productservice/AbstractIT.java
@@ -1,50 +1,42 @@
-package com.mikhailkarpov.eshop.orders.config;
+package com.mikhailkarpov.eshop.productservice;
 
+import com.mikhailkarpov.eshop.productservice.config.OrderMessagingProperties;
 import dasniko.testcontainers.keycloak.KeycloakContainer;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.amqp.rabbit.core.RabbitAdmin;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.CacheManager;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.containers.PostgreSQLContainer;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public abstract class AbstractIT {
+public class AbstractIT extends DatabaseIT {
 
-    static final PostgreSQLContainer POSTGRES;
+    static final GenericContainer REDIS;
     static final GenericContainer RABBITMQ;
     static final KeycloakContainer KEYCLOAK;
 
     static {
-        POSTGRES = (PostgreSQLContainer) new PostgreSQLContainer("postgres")
-                .withDatabaseName("order_service")
-                .withUsername("order_service")
-                .withPassword("password")
+        REDIS = new GenericContainer("redis")
+                .withExposedPorts(6379)
                 .withReuse(true);
 
         RABBITMQ = new GenericContainer("rabbitmq:3-management")
                 .withExposedPorts(5672, 15672)
                 .withReuse(true);
 
-        KEYCLOAK = new KeycloakContainer("jboss/keycloak")
-                .withRealmImportFile("/order-realm.json")
+        KEYCLOAK = new KeycloakContainer()
+                .withRealmImportFile("/product_realm.json")
                 .withReuse(true);
 
-        POSTGRES.start();
+        REDIS.start();
         RABBITMQ.start();
         KEYCLOAK.start();
-    }
-
-    @DynamicPropertySource
-    static void configDatasource(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", POSTGRES::getJdbcUrl);
-        registry.add("spring.datasource.username", POSTGRES::getUsername);
-        registry.add("spring.datasource.password", POSTGRES::getPassword);
     }
 
     @DynamicPropertySource
@@ -54,9 +46,22 @@ public abstract class AbstractIT {
     }
 
     @DynamicPropertySource
+    private static void configureRedis(DynamicPropertyRegistry registry) {
+        registry.add("spring.redis.host", REDIS::getHost);
+        registry.add("spring.redis.port", REDIS::getFirstMappedPort);
+    }
+
+    @DynamicPropertySource
     private static void configureKeycloak(DynamicPropertyRegistry registry) {
         registry.add("app.security.oauth2.jwt.issuer-uri", AbstractIT::jwtIssuerUri);
     }
+
+    private static String jwtIssuerUri() {
+        return String.format("%s/realms/product", KEYCLOAK.getAuthServerUrl());
+    }
+
+    @Autowired
+    private CacheManager cacheManager;
 
     @Autowired
     private RabbitAdmin rabbitAdmin;
@@ -71,41 +76,43 @@ public abstract class AbstractIT {
         rabbitAdmin.purgeQueue(messagingProperties.getCreatedQueue());
     }
 
-    private static String jwtIssuerUri() {
-        return String.format("%s/realms/order", KEYCLOAK.getAuthServerUrl());
+    @BeforeEach
+    void invalidateCache() {
+        cacheManager.getCacheNames().forEach(cache -> cacheManager.getCache(cache).invalidate());
     }
 
-    protected String obtainUserAccessToken() {
-        return obtainAccessToken("user", "user-pass");
+
+    protected String obtainAccessToken() {
+        return obtainAccessToken(null);
     }
 
-    protected String obtainAdminAccessToken() {
-        return obtainAccessToken("admin", "admin-pass");
-    }
+    protected String obtainAccessToken(String scope) {
 
-    private String obtainAccessToken(String username, String password) {
-
-        String clientId = "order-service";
-        String clientSecret = "67715df9-d9f4-42b5-bc81-f58868ef0149";
+        String clientId = "product-service";
+        String clientSecret = "d6b910e5-972d-4191-9b2c-57329d89864c";
+        String authUrl = jwtIssuerUri();
 
         Map<String, String> params = new HashMap<>();
         params.put("client_id", clientId);
         params.put("client_secret", clientSecret);
         params.put("grant_type", "password");
-        params.put("username", username);
-        params.put("password", password);
+        params.put("username", "user");
+        params.put("password", "user-pass");
+
+        if (scope != null) {
+            params.put("scope", scope);
+        }
 
         //@formatter:off
         return RestAssured
                 .given()
                     .contentType("application/x-www-form-urlencoded")
-                .   formParams(params)
+                    .formParams(params)
                 .when()
-                    .post(jwtIssuerUri() + "/protocol/openid-connect/token")
+                    .post(authUrl + "/protocol/openid-connect/token")
                 .then()
                     .assertThat().statusCode(200)
-                .extract()
-                    .path("access_token");
+                    .extract().path("access_token");
         //@formatter:on
     }
 }
