+++ product-service/src/test/java/com/mikhailkarpov/eshop/productservice/CacheManagerIT.java
@@ -1,4 +1,4 @@
-package com.mikhailkarpov.eshop.productservice.config;
+package com.mikhailkarpov.eshop.productservice;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
