+++ product-service/src/test/java/com/mikhailkarpov/eshop/productservice/ProductServiceApplicationIT.java
@@ -1,6 +1,5 @@
 package com.mikhailkarpov.eshop.productservice;
 
-import com.mikhailkarpov.eshop.productservice.config.AbstractIT;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
